improve response codes

+++ users-api/src/userController.ts
@@ -13,19 +13,23 @@ const addUser = (user: UserPost): User => {
 
 const deleteUser = (userId: number): User => {
   const user = findUser(userId);
-  userRepository.deleteUserById(userId);
-  return user;
+  if (user !== undefined) {
+    userRepository.deleteUserById(userId);
+    return user;
+  }
 };
 
 const findUser = (userId: number): User => userRepository.getUserById(userId);
 
 const updateUser = (userId: number, toUpdate: any): User => {
   let user = findUser(userId) as any;
-  Object.keys(toUpdate)
-    .filter((key) => Object.keys(user).includes(key))
-    .every((key) => (user[key] = toUpdate[key]));
-  userRepository.updateUser(user);
-  return user as User;
+  if (user !== undefined) {
+    Object.keys(toUpdate)
+      .filter((key) => Object.keys(user).includes(key))
+      .every((key) => (user[key] = toUpdate[key]));
+    userRepository.updateUser(user);
+    return user as User;
+  }
 };
 
 export { listUsers, addUser, deleteUser, findUser, updateUser };
